Module Tezos_raw_protocol_012_Psithaca.Baking

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Insufficient_endorsing_power of {

endorsing_power : int;

consensus_threshold : int;

}

val endorsing_rights : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * Alpha_context.Slot.t list Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.Map.t) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

For a given level computes who has the right to include an endorsement in the next block. It returns a mapping from the delegates with such rights to their endorsing slots. This function is only used by the 'validators' RPC.

val endorsing_rights_by_first_slot : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * (Alpha_context.public_key * Alpha_context.public_key_hash * int) Alpha_context.Slot.Map.t) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

Computes the endorsing rights for a given level. Returns a map from allocated first slots to their owner's public key, public key hash, and endorsing power.

val bonus_baking_reward : Alpha_context.context -> endorsing_power:int -> Alpha_context.Tez.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Computes the bonus baking reward depending on the endorsing power.

val baking_rights : Alpha_context.context -> Alpha_context.Level.t -> Alpha_context.public_key Misc.lazy_list

baking_rights ctxt level is the lazy list of contract's public key hashes that are allowed to propose for level at each round.