Module Block_header_repr.Proof_of_work

Checks if the header that would be built from the given components is valid for the given difficulty. The signature is not passed as it is does not impact the proof-of-work stamp. The stamp is checked on the hash of a block header whose signature has been zeroed-out.

val check_hash : Tezos_protocol_environment_012_Psithaca.Block_hash.t -> int64 -> bool

val check_header_proof_of_work_stamp : shell_header -> contents -> int64 -> bool

val check_proof_of_work_stamp : proof_of_work_threshold:int64 -> block_header -> unit Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult