Module Tezos_raw_protocol_012_Psithaca.Contract_manager_storage

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Unrevealed_manager_key of Contract_repr.t

| Inconsistent_hash of {

public_key : Tezos_protocol_environment_012_Psithaca.Signature.Public_key.t;

expected_hash : Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.t;

provided_hash : Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.t;

}

| Previously_revealed_key of Contract_repr.t

val init : Raw_context.t -> Contract_repr.t -> Manager_repr.manager_key -> Raw_context.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

init ctxt contract manager associates manager to contract. This function is undefined if contract has already a manager associated to it.

val is_manager_key_revealed : Raw_context.t -> Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

val reveal_manager_key : Raw_context.t -> Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.t -> Tezos_protocol_environment_012_Psithaca.Signature.Public_key.t -> Raw_context.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

val get_manager_key : ?error:Tezos_protocol_environment_012_Psithaca.Error_monad.error -> Raw_context.t -> Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.t -> Tezos_protocol_environment_012_Psithaca.Signature.Public_key.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

get_manager_key ?error ctxt pkh returns the revealed manager key of the contract represented by pkh. When error is not provided this function fails with "get_manager_key" error if pkh does not have a manager, and with Unrevealed_manager_key error if the manager has not revealed its key. When error is provided, the function fails with the provided error in both cases.

val remove_existing : Raw_context.t -> Contract_repr.t -> Raw_context.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t