Module Tezos_raw_protocol_012_Psithaca.Gas_monad

type ('a, 'trace) t

This monad combines:

a state monad where the state is the context

two levels of error monad to distinguish gas exhaustion from other errors

It is useful for backtracking on type checking errors without backtracking the consumed gas.

type ('a, 'trace) gas_monad = ('a, 'trace) t

Alias of ('a, 'trace) t to avoid confusion when the module is open

val return : 'a -> ('a, 'trace) t

monadic return operator of the gas monad

val (>>$) : ('a, 'trace) t -> ('a -> ('b, 'trace) t) -> ('b, 'trace) t

Binding operator for the gas monad

val (>|$) : ('a, 'trace) t -> ('a -> 'b) -> ('b, 'trace) t

Mapping operator for the gas monad, m >|$ f is equivalent to m >>$ fun x -> return (f x)

val (>?$) : ('a, 'trace) t -> ('a -> ('b, 'trace) Tezos_protocol_environment_012_Psithaca.Pervasives.result) -> ('b, 'trace) t

Variant of ( >>$ ) to bind uncarbonated functions

val (>??$) : ('a, 'trace) t -> (('a, 'trace) Tezos_protocol_environment_012_Psithaca.Pervasives.result -> ('b, 'trace) t) -> ('b, 'trace) t

Another variant of ( >>$ ) that lets recover from inner errors

val of_result : ('a, 'trace) Tezos_protocol_environment_012_Psithaca.Pervasives.result -> ('a, 'trace) t

gas-free embedding of tzresult values. of_result x is equivalent to return () >?$ fun () -> x

val consume_gas : Alpha_context.Gas.cost -> (unit, 'trace) t

A wrapper around Gas.consume. If that fails, the whole computation within the Gas_monad returns an error. See the Alpha_context.Gas module for details.

val run : Alpha_context.context -> ('a, 'trace) t -> (('a, 'trace) Tezos_protocol_environment_012_Psithaca.Pervasives.result * Alpha_context.context) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Escaping the gas monad

val record_trace_eval : (unit -> 'err) -> ('a, 'err Tezos_protocol_environment_012_Psithaca.Error_monad.trace) t -> ('a, 'err Tezos_protocol_environment_012_Psithaca.Error_monad.trace) t

re-export of Error_monad.record_trace_eval. This function has no effect in the case of a gas-exhaustion error.