Module Tezos_raw_protocol_012_Psithaca.Raw_context

Errors

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Too_many_internal_operations

type missing_key_kind =

| Get

| Set

| Del

| Copy

type storage_error =

| Incompatible_protocol_version of string

| Missing_key of string list * missing_key_kind

| Existing_key of string list

| Corrupted_data of string list

An internal storage error that should not happen

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Storage_error of storage_error

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Failed_to_parse_parameter of bytes

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Failed_to_decode_parameter of Tezos_protocol_environment_012_Psithaca.Data_encoding.json * string

val storage_error : storage_error -> 'a Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Abstract Context

type t

Abstract view of the context. Includes a handle to the functional key-value database (Context.t) along with some in-memory values (gas, etc.).

type root = t

val prepare : level:Tezos_protocol_environment_012_Psithaca.Int32.t -> predecessor_timestamp:Tezos_protocol_environment_012_Psithaca.Time.t -> timestamp:Tezos_protocol_environment_012_Psithaca.Time.t -> Tezos_protocol_environment_012_Psithaca.Context.t -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

Retrieves the state of the database and gives its abstract view. It also returns wether this is the first block validated with this version of the protocol.

type previous_protocol =

| Genesis of Parameters_repr.t

| Hangzhou_011

val prepare_first_block : level:int32 -> timestamp:Tezos_protocol_environment_012_Psithaca.Time.t -> Tezos_protocol_environment_012_Psithaca.Context.t -> (previous_protocol * t) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

val activate : t -> Tezos_protocol_environment_012_Psithaca.Protocol_hash.t -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

val recover : t -> Tezos_protocol_environment_012_Psithaca.Context.t

Returns the state of the database resulting of operations on its abstract view

val current_level : t -> Level_repr.t

val predecessor_timestamp : t -> Tezos_protocol_environment_012_Psithaca.Time.t

val current_timestamp : t -> Tezos_protocol_environment_012_Psithaca.Time.t

val constants : t -> Constants_repr.parametric

val patch_constants : t -> (Constants_repr.parametric -> Constants_repr.parametric) -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

val round_durations : t -> Round_repr.Durations.t

val cycle_eras : t -> Level_repr.cycle_eras

Retrieve the cycle eras.

val credit_collected_fees_only_call_from_token : t -> Tez_repr.t -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Increment the current block fee stash that will be credited to the payload producer's account at finalize_application

val spend_collected_fees_only_call_from_token : t -> Tez_repr.t -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Decrement the current block fee stash that will be credited to the payload producer's account at finalize_application

val get_collected_fees : t -> Tez_repr.t

Returns the current block fee stash that will be credited to the payload producer's account at finalize_application

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Gas_limit_too_high

val check_gas_limit_is_valid : t -> 'a Gas_limit_repr.Arith.t -> unit Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val consume_gas_limit_in_block : t -> 'a Gas_limit_repr.Arith.t -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val set_gas_limit : t -> 'a Gas_limit_repr.Arith.t -> t

val set_gas_unlimited : t -> t

val gas_level : t -> Gas_limit_repr.t

val gas_consumed : since:t -> until:t -> Gas_limit_repr.Arith.fp

val remaining_operation_gas : t -> Gas_limit_repr.Arith.fp

val update_remaining_operation_gas : t -> Gas_limit_repr.Arith.fp -> t

val block_gas_level : t -> Gas_limit_repr.Arith.fp

val update_remaining_block_gas : t -> Gas_limit_repr.Arith.fp -> t

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Undefined_operation_nonce

val init_origination_nonce : t -> Tezos_protocol_environment_012_Psithaca.Operation_hash.t -> t

val get_origination_nonce : t -> Contract_repr.origination_nonce Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val increment_origination_nonce : t -> (t * Contract_repr.origination_nonce) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val unset_origination_nonce : t -> t

Generic accessors

type key = string list

type value = bytes

type tree

module type T = Raw_context_intf.T with type root := root and type key := key and type value := value and type tree := tree

include T with type t := t

include Raw_context_intf.VIEW with type key := key with type value := value with type tree := tree with type t := t

Getters

val mem : t -> key -> bool Tezos_protocol_environment_012_Psithaca.Lwt.t

mem t k is an Lwt promise that resolves to true iff k is bound to a value in t.

val mem_tree : t -> key -> bool Tezos_protocol_environment_012_Psithaca.Lwt.t

mem_tree t k is like mem but for trees.

val get : t -> key -> value Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

get t k is an Lwt promise that resolves to Ok v if k is bound to the value v in t and Storage_ErrorMissing_key otherwise.

val get_tree : t -> key -> tree Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

get_tree is like get but for trees.

val find : t -> key -> value option Tezos_protocol_environment_012_Psithaca.Lwt.t

find t k is an Lwt promise that resolves to Some v if k is bound to the value v in t and None otherwise.

val find_tree : t -> key -> tree option Tezos_protocol_environment_012_Psithaca.Lwt.t

find_tree t k is like find but for trees.

val list : t -> ?offset:int -> ?length:int -> key -> (string * tree) list Tezos_protocol_environment_012_Psithaca.Lwt.t

list t key is the list of files and sub-nodes stored under k in t. The result order is not specified but is stable.

offset and length are used for pagination.

Setters

val init : t -> key -> value -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

init t k v is an Lwt promise that resolves to Ok c if:

k is unbound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_errorExisting_key if k is already bound in t.

val init_tree : t -> key -> tree -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

init_tree is like init but for trees.

val update : t -> key -> value -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

update t k v is an Lwt promise that resolves to Ok c if:

k is bound in t;

k is bound to v in c;

and c is similar to t otherwise.

It is Storage_errorMissing_key if k is not already bound in t.

val update_tree : t -> key -> tree -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

update_tree is like update but for trees.

val add : t -> key -> value -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

add t k v is an Lwt promise that resolves to c such that:

k is bound to v in c;

and c is similar to t otherwise.

If k was already bound in t to a value that is physically equal to v, the result of the function is a promise that resolves to t. Otherwise, the previous binding of k in t disappears.

val add_tree : t -> key -> tree -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

add_tree is like add but for trees.

val remove : t -> key -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

remove t k v is an Lwt promise that resolves to c such that:

k is unbound in c;

and c is similar to t otherwise.

val remove_existing : t -> key -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

remove_existing t k v is an Lwt promise that resolves to Ok c if:

k is bound in t to a value;

k is unbound in c;

and c is similar to t otherwise.

val remove_existing_tree : t -> key -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

remove_existing_tree t k v is an Lwt promise that reolves to Ok c if:

k is bound in t to a tree;

k is unbound in c;

and c is similar to t otherwise.

val add_or_remove : t -> key -> value option -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

add_or_remove t k v is:

add t k x if v is Some x;

remove t k otherwise.

val add_or_remove_tree : t -> key -> tree option -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

add_or_remove_tree t k v is:

add_tree t k x if v is Some x;

remove t k otherwise.

Folds

val fold : ?depth:[ `Eq of int | `Le of int | `Lt of int | `Ge of int | `Gt of int ] -> t -> key -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(key -> tree -> 'a -> 'a Tezos_protocol_environment_012_Psithaca.Lwt.t) -> 'a Tezos_protocol_environment_012_Psithaca.Lwt.t

fold ?depth t root ~init ~f recursively folds over the trees and values of t. The f callbacks are called with a key relative to root. f is never called with an empty key for values; i.e., folding over a value is a no-op.

Elements are traversed in lexical order of keys.

The depth is 0-indexed. If depth is set (by default it is not), then f is only called when the conditions described by the parameter is true:

Eq d folds over nodes and contents of depth exactly d.

Lt d folds over nodes and contents of depth strictly less than d.

Le d folds over nodes and contents of depth less than or equal to d.

Gt d folds over nodes and contents of depth strictly more than d.

Ge d folds over nodes and contents of depth more than or equal to d.

module Tree : Raw_context_intf.TREE with type t := t and type key := key and type value := value and type tree := tree

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

val project : t -> root

Internally used in Storage_functors to escape from a view.

val absolute_key : t -> key -> key

Internally used in Storage_functors to retrieve a full key from partial key relative a view.

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Block_quota_exceeded

Raised if block gas quota is exhausted during gas consumption.

type Tezos_protocol_environment_012_Psithaca.Error_monad.error +=

| Operation_quota_exceeded

Raised if operation gas quota is exhausted during gas consumption.

val consume_gas : t -> Gas_limit_repr.cost -> t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Internally used in Storage_functors to consume gas from within a view. May raise Block_quota_exceeded or Operation_quota_exceeded.

val check_enough_gas : t -> Gas_limit_repr.cost -> unit Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Check if consume_gas will fail

val description : t Storage_description.t

val reset_internal_nonce : t -> t

Initialize the local nonce used for preventing a script to duplicate an internal operation to replay it.

val fresh_internal_nonce : t -> (t * int) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

Increments the internal operation nonce.

val record_internal_nonce : t -> int -> t

Mark an internal operation nonce as taken.

val internal_nonce_already_recorded : t -> int -> bool

Check is the internal operation nonce has been taken.

val fold_map_temporary_lazy_storage_ids : t -> (Lazy_storage_kind.Temp_ids.t -> Lazy_storage_kind.Temp_ids.t * 'res) -> t * 'res

val map_temporary_lazy_storage_ids_s : t -> (Lazy_storage_kind.Temp_ids.t -> (t * Lazy_storage_kind.Temp_ids.t) Tezos_protocol_environment_012_Psithaca.Lwt.t) -> t Tezos_protocol_environment_012_Psithaca.Lwt.t

module Cache : Tezos_protocol_environment_012_Psithaca.Context.CACHE with type t := t and type size := int and type index := int and type identifier := string and type key = Tezos_protocol_environment_012_Psithaca.Context.Cache.key and type value = Tezos_protocol_environment_012_Psithaca.Context.Cache.value

val record_non_consensus_operation_hash : t -> Tezos_protocol_environment_012_Psithaca.Operation_hash.t -> t

val non_consensus_operations : t -> Tezos_protocol_environment_012_Psithaca.Operation_hash.t list

val set_sampler_for_cycle : t -> Cycle_repr.t -> (Seed_repr.seed * (Tezos_protocol_environment_012_Psithaca.Signature.public_key * Tezos_protocol_environment_012_Psithaca.Signature.public_key_hash) Sampler.t) -> (t, [ `Sampler_already_set ]) Tezos_protocol_environment_012_Psithaca.Pervasives.result

set_sampler_for_cycle ctxt cycle sampler evaluates to Ok c with c verifying sampler_for_cycle c cycle = sampler if no sampler was set for the same cycle beforehand. In the other case, it returns Error `Sampler_already_set.

val sampler_for_cycle : t -> Cycle_repr.t -> (Seed_repr.seed * (Tezos_protocol_environment_012_Psithaca.Signature.public_key * Tezos_protocol_environment_012_Psithaca.Signature.public_key_hash) Sampler.t, [ `Sampler_not_set ]) Tezos_protocol_environment_012_Psithaca.Pervasives.result

sampler_for_cycle ctxt cycle evaluates to Ok sampler if a sampler was set for cycle using set_sampler_for_cycle. Otherwise, it returns Error `Sampler_not_set.

val stake_distribution_for_current_cycle : t -> Tez_repr.t Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.Map.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val init_stake_distribution_for_current_cycle : t -> Tez_repr.t Tezos_protocol_environment_012_Psithaca.Signature.Public_key_hash.Map.t -> t

module type CONSENSUS = sig ... end

module Consensus : CONSENSUS with type t := t and type slot := Slot_repr.t and type 'a slot_map := 'a Slot_repr.Map.t and type slot_set := Slot_repr.Set.t and type round := Round_repr.t