Module type Sampler.Mass

Mass is the module type describing the measure associated to points.

type t

t is the type describing the measure associated to points.

val encoding : t Tezos_protocol_environment_012_Psithaca.Data_encoding.t

val zero : t

val of_int : int -> t

val mul : t -> t -> t

val add : t -> t -> t

val sub : t -> t -> t

val (=) : t -> t -> bool

val (<=) : t -> t -> bool

val (<) : t -> t -> bool