Module Script_interpreter.Internals

type local_gas_counter = int

Internally, the interpretation loop uses a local gas counter.

type outdated_context =

| OutDatedContext of Alpha_context.context

During the evaluation, the gas level in the context is outdated. See comments in the implementation file for more details.

val next : Script_typed_ir.logger option -> (outdated_context * step_constants) -> local_gas_counter -> ('a, 's, 'r, 'f) Script_typed_ir.continuation -> 'a -> 's -> ('r * 'f * outdated_context * local_gas_counter) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

next logger (ctxt, step_constants) local_gas_counter ks accu stack is an internal function which interprets the continuation ks to execute the interpreter on the current A-stack.

val step : (outdated_context * step_constants) -> local_gas_counter -> ('a, 's, 'r, 'f) Script_typed_ir.kinstr -> 'a -> 's -> ('r * 'f * outdated_context * local_gas_counter) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t