Module Tezos_raw_protocol_012_Psithaca.Script_typed_ir

type step_constants = {

source : Alpha_context.Contract.t;

payer : Alpha_context.Contract.t;

self : Alpha_context.Contract.t;

amount : Alpha_context.Tez.t;

chain_id : Tezos_protocol_environment_012_Psithaca.Chain_id.t;

now : Alpha_context.Script_timestamp.t;

level : Alpha_context.Script_int.n Alpha_context.Script_int.num;

}

type never = |

type address = Alpha_context.Contract.t * string

type ('a, 'b) pair = 'a * 'b

type ('a, 'b) union =

| L of 'a

| R of 'b

type operation = Alpha_context.packed_internal_operation * Alpha_context.Lazy_storage.diffs option

type 'a ticket = {

ticketer : Alpha_context.Contract.t;

contents : 'a;

amount : Alpha_context.Script_int.n Alpha_context.Script_int.num;

}

type empty_cell =

| EmptyCell

type end_of_stack = empty_cell * empty_cell

module Type_size : sig ... end

type 'a ty_metadata = {

annot : Script_ir_annot.type_annot option;

size : 'a Type_size.t;

}

type _ comparable_ty =

| Unit_key : unit ty_metadata -> unit comparable_ty

| Never_key : never ty_metadata -> never comparable_ty

| Int_key : Alpha_context.Script_int.z Alpha_context.Script_int.num ty_metadata -> Alpha_context.Script_int.z Alpha_context.Script_int.num comparable_ty

| Nat_key : Alpha_context.Script_int.n Alpha_context.Script_int.num ty_metadata -> Alpha_context.Script_int.n Alpha_context.Script_int.num comparable_ty

| Signature_key : Alpha_context.signature ty_metadata -> Alpha_context.signature comparable_ty

| String_key : Alpha_context.Script_string.t ty_metadata -> Alpha_context.Script_string.t comparable_ty

| Bytes_key : Tezos_protocol_environment_012_Psithaca.Bytes.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Bytes.t comparable_ty

| Mutez_key : Alpha_context.Tez.t ty_metadata -> Alpha_context.Tez.t comparable_ty

| Bool_key : bool ty_metadata -> bool comparable_ty

| Key_hash_key : Alpha_context.public_key_hash ty_metadata -> Alpha_context.public_key_hash comparable_ty

| Key_key : Alpha_context.public_key ty_metadata -> Alpha_context.public_key comparable_ty

| Timestamp_key : Alpha_context.Script_timestamp.t ty_metadata -> Alpha_context.Script_timestamp.t comparable_ty

| Chain_id_key : Tezos_protocol_environment_012_Psithaca.Chain_id.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Chain_id.t comparable_ty

| Address_key : address ty_metadata -> address comparable_ty

| Pair_key : ('a comparable_ty * Script_ir_annot.field_annot option) * ('b comparable_ty * Script_ir_annot.field_annot option) * ('a, 'b) pair ty_metadata -> ('a, 'b) pair comparable_ty

| Union_key : ('a comparable_ty * Script_ir_annot.field_annot option) * ('b comparable_ty * Script_ir_annot.field_annot option) * ('a, 'b) union ty_metadata -> ('a, 'b) union comparable_ty

| Option_key : 'v comparable_ty * 'v option ty_metadata -> 'v option comparable_ty

val unit_key : annot:Script_ir_annot.type_annot option -> unit comparable_ty

val never_key : annot:Script_ir_annot.type_annot option -> never comparable_ty

val int_key : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_int.z Alpha_context.Script_int.num comparable_ty

val nat_key : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_int.n Alpha_context.Script_int.num comparable_ty

val signature_key : annot:Script_ir_annot.type_annot option -> Alpha_context.signature comparable_ty

val string_key : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_string.t comparable_ty

val bytes_key : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Bytes.t comparable_ty

val mutez_key : annot:Script_ir_annot.type_annot option -> Alpha_context.Tez.t comparable_ty

val bool_key : annot:Script_ir_annot.type_annot option -> bool comparable_ty

val key_hash_key : annot:Script_ir_annot.type_annot option -> Alpha_context.public_key_hash comparable_ty

val key_key : annot:Script_ir_annot.type_annot option -> Alpha_context.public_key comparable_ty

val timestamp_key : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_timestamp.t comparable_ty

val chain_id_key : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Chain_id.t comparable_ty

val address_key : annot:Script_ir_annot.type_annot option -> address comparable_ty

val pair_key : Alpha_context.Script.location -> ('a comparable_ty * Script_ir_annot.field_annot option) -> ('b comparable_ty * Script_ir_annot.field_annot option) -> annot:Script_ir_annot.type_annot option -> ('a, 'b) pair comparable_ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val pair_3_key : Alpha_context.Script.location -> ('a comparable_ty * Script_ir_annot.field_annot option) -> ('b comparable_ty * Script_ir_annot.field_annot option) -> ('c comparable_ty * Script_ir_annot.field_annot option) -> ('a, ('b, 'c) pair) pair comparable_ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val union_key : Alpha_context.Script.location -> ('a comparable_ty * Script_ir_annot.field_annot option) -> ('b comparable_ty * Script_ir_annot.field_annot option) -> annot:Script_ir_annot.type_annot option -> ('a, 'b) union comparable_ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val option_key : Alpha_context.Script.location -> 'v comparable_ty -> annot:Script_ir_annot.type_annot option -> 'v option comparable_ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

module type Boxed_set_OPS = sig ... end

module type Boxed_set = sig ... end

type 'elt set = (module Boxed_set with type elt = 'elt)

module type Boxed_map_OPS = sig ... end

module type Boxed_map = sig ... end

type ('key, 'value) map = (module Boxed_map with type key = 'key and type value = 'value)

module Big_map_overlay : Tezos_protocol_environment_012_Psithaca.Map.S with type key = Script_expr_hash.t

type ('key, 'value) big_map_overlay = {

map : ('key * 'value option) Big_map_overlay.t;

size : int;

}

type 'elt boxed_list = {

elements : 'elt list;

length : int;

}

module SMap : Tezos_protocol_environment_012_Psithaca.Map.S with type key = Alpha_context.Script_string.t

type view = {

input_ty : Alpha_context.Script.node;

output_ty : Alpha_context.Script.node;

view_code : Alpha_context.Script.node;

}

type ('arg, 'storage) script = {

code : (('arg, 'storage) pair, (operation boxed_list, 'storage) pair) lambda;

arg_type : 'arg ty;

storage : 'storage;

storage_type : 'storage ty;

views : view SMap.t;

root_name : Script_ir_annot.field_annot option;

code_size : Cache_memory_helpers.sint;

}

and ('before_top, 'before, 'result_top, 'result) kinstr =

| IDrop : ('a, 'b * 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDup : ('a, 's) kinfo * ('a, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISwap : ('a, 'b * 's) kinfo * ('b, 'a * 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IConst : ('a, 's) kinfo * 'ty * ('ty, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICons_pair : ('a, 'b * 's) kinfo * ('a * 'b, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| ICar : ('a * 'b, 's) kinfo * ('a, 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| ICdr : ('a * 'b, 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| IUnpair : ('a * 'b, 's) kinfo * ('a, 'b * 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| ICons_some : ('v, 's) kinfo * ('v option, 's, 'r, 'f) kinstr -> ('v, 's, 'r, 'f) kinstr

| ICons_none : ('a, 's) kinfo * ('b option, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IIf_none : {

kinfo : ('a option, 'b * 's) kinfo;

branch_if_none : ('b, 's, 'c, 't) kinstr;

branch_if_some : ('a, 'b * 's, 'c, 't) kinstr;

k : ('c, 't, 'r, 'f) kinstr;

} -> ('a option, 'b * 's, 'r, 'f) kinstr

| IOpt_map : {

kinfo : ('a option, 's) kinfo;

body : ('a, 's, 'b, 's) kinstr;

k : ('b option, 's, 'c, 't) kinstr;

} -> ('a option, 's, 'c, 't) kinstr

| ICons_left : ('a, 's) kinfo * (('a, 'b) union, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICons_right : ('b, 's) kinfo * (('a, 'b) union, 's, 'r, 'f) kinstr -> ('b, 's, 'r, 'f) kinstr

| IIf_left : {

kinfo : (('a, 'b) union, 's) kinfo;

branch_if_left : ('a, 's, 'c, 't) kinstr;

branch_if_right : ('b, 's, 'c, 't) kinstr;

k : ('c, 't, 'r, 'f) kinstr;

} -> (('a, 'b) union, 's, 'r, 'f) kinstr

| ICons_list : ('a, 'a boxed_list * 's) kinfo * ('a boxed_list, 's, 'r, 'f) kinstr -> ('a, 'a boxed_list * 's, 'r, 'f) kinstr

| INil : ('a, 's) kinfo * ('b boxed_list, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IIf_cons : {

kinfo : ('a boxed_list, 'b * 's) kinfo;

branch_if_cons : ('a, 'a boxed_list * ('b * 's), 'c, 't) kinstr;

branch_if_nil : ('b, 's, 'c, 't) kinstr;

k : ('c, 't, 'r, 'f) kinstr;

} -> ('a boxed_list, 'b * 's, 'r, 'f) kinstr

| IList_map : ('a boxed_list, 'c * 's) kinfo * ('a, 'c * 's, 'b, 'c * 's) kinstr * ('b boxed_list, 'c * 's, 'r, 'f) kinstr -> ('a boxed_list, 'c * 's, 'r, 'f) kinstr

| IList_iter : ('a boxed_list, 'b * 's) kinfo * ('a, 'b * 's, 'b, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> ('a boxed_list, 'b * 's, 'r, 'f) kinstr

| IList_size : ('a boxed_list, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a boxed_list, 's, 'r, 'f) kinstr

| IEmpty_set : ('a, 's) kinfo * 'b comparable_ty * ('b set, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISet_iter : ('a set, 'b * 's) kinfo * ('a, 'b * 's, 'b, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> ('a set, 'b * 's, 'r, 'f) kinstr

| ISet_mem : ('a, 'a set * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, 'a set * 's, 'r, 'f) kinstr

| ISet_update : ('a, bool * ('a set * 's)) kinfo * ('a set, 's, 'r, 'f) kinstr -> ('a, bool * ('a set * 's), 'r, 'f) kinstr

| ISet_size : ('a set, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a set, 's, 'r, 'f) kinstr

| IEmpty_map : ('a, 's) kinfo * 'b comparable_ty * (('b, 'c) map, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IMap_map : (('a, 'b) map, 'd * 's) kinfo * ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * (('a, 'c) map, 'd * 's, 'r, 'f) kinstr -> (('a, 'b) map, 'd * 's, 'r, 'f) kinstr

| IMap_iter : (('a, 'b) map, 'c * 's) kinfo * ('a * 'b, 'c * 's, 'c, 's) kinstr * ('c, 's, 'r, 'f) kinstr -> (('a, 'b) map, 'c * 's, 'r, 'f) kinstr

| IMap_mem : ('a, ('a, 'b) map * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) map * 's, 'r, 'f) kinstr

| IMap_get : ('a, ('a, 'b) map * 's) kinfo * ('b option, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) map * 's, 'r, 'f) kinstr

| IMap_update : ('a, 'b option * (('a, 'b) map * 's)) kinfo * (('a, 'b) map, 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) map * 's), 'r, 'f) kinstr

| IMap_get_and_update : ('a, 'b option * (('a, 'b) map * 's)) kinfo * ('b option, ('a, 'b) map * 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) map * 's), 'r, 'f) kinstr

| IMap_size : (('a, 'b) map, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (('a, 'b) map, 's, 'r, 'f) kinstr

| IEmpty_big_map : ('a, 's) kinfo * 'b comparable_ty * 'c ty * (('b, 'c) big_map, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IBig_map_mem : ('a, ('a, 'b) big_map * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) big_map * 's, 'r, 'f) kinstr

| IBig_map_get : ('a, ('a, 'b) big_map * 's) kinfo * ('b option, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) big_map * 's, 'r, 'f) kinstr

| IBig_map_update : ('a, 'b option * (('a, 'b) big_map * 's)) kinfo * (('a, 'b) big_map, 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) big_map * 's), 'r, 'f) kinstr

| IBig_map_get_and_update : ('a, 'b option * (('a, 'b) big_map * 's)) kinfo * ('b option, ('a, 'b) big_map * 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) big_map * 's), 'r, 'f) kinstr

| IConcat_string : (Alpha_context.Script_string.t boxed_list, 's) kinfo * (Alpha_context.Script_string.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_string.t boxed_list, 's, 'r, 'f) kinstr

| IConcat_string_pair : (Alpha_context.Script_string.t, Alpha_context.Script_string.t * 's) kinfo * (Alpha_context.Script_string.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_string.t, Alpha_context.Script_string.t * 's, 'r, 'f) kinstr

| ISlice_string : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (Alpha_context.Script_string.t * 's)) kinfo * (Alpha_context.Script_string.t option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * ( Alpha_context.Script_string.t * 's), 'r, 'f) kinstr

| IString_size : (Alpha_context.Script_string.t, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_string.t, 's, 'r, 'f) kinstr

| IConcat_bytes : (bytes boxed_list, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes boxed_list, 's, 'r, 'f) kinstr

| IConcat_bytes_pair : (bytes, bytes * 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, bytes * 's, 'r, 'f) kinstr

| ISlice_bytes : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (bytes * 's)) kinfo * (bytes option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (bytes * 's), 'r, 'f) kinstr

| IBytes_size : (bytes, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IAdd_seconds_to_timestamp : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_timestamp.t * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_timestamp.t * 's, 'r, 'f) kinstr

| IAdd_timestamp_to_seconds : (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ISub_timestamp_seconds : (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IDiff_timestamps : (Alpha_context.Script_timestamp.t, Alpha_context.Script_timestamp.t * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_timestamp.t * 's, 'r, 'f) kinstr

| IAdd_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| ISub_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t option, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| ISub_tez_legacy : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IMul_teznat : (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_nattez : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IEdiv_teznat : (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Tez.t, Alpha_context.Tez.t) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * ((Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IOr : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| IAnd : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| IXor : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| INot : (bool, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, 's, 'r, 'f) kinstr

| IIs_nat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INeg : ('a Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAbs_int : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IInt_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAdd_int : ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAdd_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ISub_int : ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_int : ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a Alpha_context.Script_int.num * 's) kinfo * ('a Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_int : ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a Alpha_context.Script_int.num * 's) kinfo * (('a Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ILsl_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ILsr_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IOr_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAnd_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAnd_int_nat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IXor_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| INot_int : ('a Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IIf : {

kinfo : (bool, 'a * 's) kinfo;

branch_if_true : ('a, 's, 'b, 'u) kinstr;

branch_if_false : ('a, 's, 'b, 'u) kinstr;

k : ('b, 'u, 'r, 'f) kinstr;

} -> (bool, 'a * 's, 'r, 'f) kinstr

| ILoop : (bool, 'a * 's) kinfo * ('a, 's, bool, 'a * 's) kinstr * ('a, 's, 'r, 'f) kinstr -> (bool, 'a * 's, 'r, 'f) kinstr

| ILoop_left : (('a, 'b) union, 's) kinfo * ('a, 's, ('a, 'b) union, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> (('a, 'b) union, 's, 'r, 'f) kinstr

| IDip : ('a, 'b * 's) kinfo * ('b, 's, 'c, 't) kinstr * ('a, 'c * 't, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IExec : ('a, ('a, 'b) lambda * 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) lambda * 's, 'r, 'f) kinstr

| IApply : ('a, ('a * 'b, 'c) lambda * 's) kinfo * 'a ty * (('b, 'c) lambda, 's, 'r, 'f) kinstr -> ('a, ('a * 'b, 'c) lambda * 's, 'r, 'f) kinstr

| ILambda : ('a, 's) kinfo * ('b, 'c) lambda * (('b, 'c) lambda, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IFailwith : ('a, 's) kinfo * Alpha_context.Script.location * 'a ty -> ('a, 's, 'r, 'f) kinstr

| ICompare : ('a, 'a * 's) kinfo * 'a comparable_ty * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a, 'a * 's, 'r, 'f) kinstr

| IEq : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INeq : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| ILt : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IGt : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| ILe : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IGe : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAddress : ('a typed_contract, 's) kinfo * (address, 's, 'r, 'f) kinstr -> ('a typed_contract, 's, 'r, 'f) kinstr

| IContract : (address, 's) kinfo * 'a ty * string * ('a typed_contract option, 's, 'r, 'f) kinstr -> (address, 's, 'r, 'f) kinstr

| IView : ('a, address * 's) kinfo * ('a, 'b) view_signature * ('b option, 's, 'r, 'f) kinstr -> ('a, address * 's, 'r, 'f) kinstr

| ITransfer_tokens : ('a, Alpha_context.Tez.t * ('a typed_contract * 's)) kinfo * (operation, 's, 'r, 'f) kinstr -> ('a, Alpha_context.Tez.t * ('a typed_contract * 's), 'r, 'f) kinstr

| IImplicit_account : (Alpha_context.public_key_hash, 's) kinfo * (unit typed_contract, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr

| ICreate_contract : {

kinfo : (Alpha_context.public_key_hash option, Alpha_context.Tez.t * ('a * 's)) kinfo;

storage_type : 'a ty;

arg_type : 'b ty;

lambda : ('b * 'a, operation boxed_list * 'a) lambda;

views : view SMap.t;

root_name : Script_ir_annot.field_annot option;

k : (operation, address * 's, 'r, 'f) kinstr;

} -> (Alpha_context.public_key_hash option, Alpha_context.Tez.t * ('a * 's), 'r, 'f) kinstr

| ISet_delegate : (Alpha_context.public_key_hash option, 's) kinfo * (operation, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash option, 's, 'r, 'f) kinstr

| INow : ('a, 's) kinfo * (Alpha_context.Script_timestamp.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IBalance : ('a, 's) kinfo * (Alpha_context.Tez.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ILevel : ('a, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICheck_signature : (Alpha_context.public_key, Alpha_context.signature * (bytes * 's)) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.public_key, Alpha_context.signature * (bytes * 's), 'r, 'f) kinstr

| IHash_key : (Alpha_context.public_key, 's) kinfo * (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr -> (Alpha_context.public_key, 's, 'r, 'f) kinstr

| IPack : ('a, 's) kinfo * 'a ty * (bytes, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IUnpack : (bytes, 's) kinfo * 'a ty * ('a option, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IBlake2b : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha256 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha512 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISource : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISender : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISelf : ('a, 's) kinfo * 'b ty * string * ('b typed_contract, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISelf_address : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IAmount : ('a, 's) kinfo * (Alpha_context.Tez.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISapling_empty_state : ('a, 's) kinfo * Alpha_context.Sapling.Memo_size.t * (Alpha_context.Sapling.state, 'a * 's, 'b, 'f) kinstr -> ('a, 's, 'b, 'f) kinstr

| ISapling_verify_update : (Alpha_context.Sapling.transaction, Alpha_context.Sapling.state * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Sapling.state) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Sapling.transaction, Alpha_context.Sapling.state * 's, 'r, 'f) kinstr

| IDig : ('a, 's) kinfo * int * ('b, 'c * 't, 'c, 't, 'a, 's, 'd, 'u) stack_prefix_preservation_witness * ('b, 'd * 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IDug : ('a, 'b * 's) kinfo * int * ('c, 't, 'a, 'c * 't, 'b, 's, 'd, 'u) stack_prefix_preservation_witness * ('d, 'u, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDipn : ('a, 's) kinfo * int * ('c, 't, 'd, 'v, 'a, 's, 'b, 'u) stack_prefix_preservation_witness * ('c, 't, 'd, 'v) kinstr * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IDropn : ('a, 's) kinfo * int * ('b, 'u, 'b, 'u, 'a, 's, 'a, 's) stack_prefix_preservation_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IChainId : ('a, 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Chain_id.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| INever : (never, 's) kinfo -> (never, 's, 'r, 'f) kinstr

| IVoting_power : (Alpha_context.public_key_hash, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr

| ITotal_voting_power : ('a, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IKeccak : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha3 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IAdd_bls12_381_g1 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t * 's, 'r, 'f) kinstr

| IAdd_bls12_381_g2 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t * 's, 'r, 'f) kinstr

| IAdd_bls12_381_fr : (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_g1 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_g2 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_fr : (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_z_fr : (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 'a Alpha_context.Script_int.num * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 'a Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_bls12_381_fr_z : ('a Alpha_context.Script_int.num, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IInt_bls12_381_fr : (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_g1 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_g2 : (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_fr : (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's) kinfo * (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t, 's, 'r, 'f) kinstr

| IPairing_check_bls12_381 : ((Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t) pair boxed_list, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ((Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t, Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t) pair boxed_list, 's, 'r, 'f) kinstr

| IComb : ('a, 's) kinfo * int * ('a * 's, 'b * 'u) comb_gadt_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IUncomb : ('a, 's) kinfo * int * ('a * 's, 'b * 'u) uncomb_gadt_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IComb_get : ('t, 's) kinfo * int * ('t, 'v) comb_get_gadt_witness * ('v, 's, 'r, 'f) kinstr -> ('t, 's, 'r, 'f) kinstr

| IComb_set : ('a, 'b * 's) kinfo * int * ('a, 'b, 'c) comb_set_gadt_witness * ('c, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDup_n : ('a, 's) kinfo * int * ('a * 's, 't) dup_n_gadt_witness * ('t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ITicket : ('a, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ('a ticket, 's, 'r, 'f) kinstr -> ('a, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IRead_ticket : ('a ticket, 's) kinfo * (address * ('a * Alpha_context.Script_int.n Alpha_context.Script_int.num), 'a ticket * 's, 'r, 'f) kinstr -> ('a ticket, 's, 'r, 'f) kinstr

| ISplit_ticket : ('a ticket, (Alpha_context.Script_int.n Alpha_context.Script_int.num * Alpha_context.Script_int.n Alpha_context.Script_int.num) * 's) kinfo * (('a ticket * 'a ticket) option, 's, 'r, 'f) kinstr -> ('a ticket, (Alpha_context.Script_int.n Alpha_context.Script_int.num * Alpha_context.Script_int.n Alpha_context.Script_int.num) * 's, 'r, 'f) kinstr

| IJoin_tickets : ('a ticket * 'a ticket, 's) kinfo * 'a comparable_ty * ('a ticket option, 's, 'r, 'f) kinstr -> ('a ticket * 'a ticket, 's, 'r, 'f) kinstr

| IOpen_chest : (Tezos_protocol_environment_012_Psithaca.Timelock.chest_key, Tezos_protocol_environment_012_Psithaca.Timelock.chest * (Alpha_context.Script_int.n Alpha_context.Script_int.num * 's)) kinfo * ((bytes, bool) union, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment_012_Psithaca.Timelock.chest_key, Tezos_protocol_environment_012_Psithaca.Timelock.chest * (Alpha_context.Script_int.n Alpha_context.Script_int.num * 's), 'r, 'f) kinstr

| IHalt : ('a, 's) kinfo -> ('a, 's, 'a, 's) kinstr

| ILog : ('a, 's) kinfo * logging_event * logger * ('a, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

and logging_event =

| LogEntry : logging_event

| LogExit : ('b, 'u) kinfo -> logging_event

and ('arg, 'ret) lambda =

| Lam : ('arg, end_of_stack, 'ret, end_of_stack) kdescr * Alpha_context.Script.node -> ('arg, 'ret) lambda

and 'arg typed_contract = 'arg ty * address

and (_, _, _, _) continuation =

| KNil : ('r, 'f, 'r, 'f) continuation

| KCons : ('a, 's, 'b, 't) kinstr * ('b, 't, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KReturn : 's * ('a, 's, 'r, 'f) continuation -> ('a, end_of_stack, 'r, 'f) continuation

| KMap_head : ('a -> 'b) * ('b, 's, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KUndip : 'b * ('b, 'a * 's, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KLoop_in : ('a, 's, bool, 'a * 's) kinstr * ('a, 's, 'r, 'f) continuation -> (bool, 'a * 's, 'r, 'f) continuation

| KLoop_in_left : ('a, 's, ('a, 'b) union, 's) kinstr * ('b, 's, 'r, 'f) continuation -> (('a, 'b) union, 's, 'r, 'f) continuation

| KIter : ('a, 'b * 's, 'b, 's) kinstr * 'a list * ('b, 's, 'r, 'f) continuation -> ('b, 's, 'r, 'f) continuation

| KList_enter_body : ('a, 'c * 's, 'b, 'c * 's) kinstr * 'a list * 'b list * int * ('b boxed_list, 'c * 's, 'r, 'f) continuation -> ('c, 's, 'r, 'f) continuation

| KList_exit_body : ('a, 'c * 's, 'b, 'c * 's) kinstr * 'a list * 'b list * int * ('b boxed_list, 'c * 's, 'r, 'f) continuation -> ('b, 'c * 's, 'r, 'f) continuation

| KMap_enter_body : ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * ('a * 'b) list * ('a, 'c) map * (('a, 'c) map, 'd * 's, 'r, 'f) continuation -> ('d, 's, 'r, 'f) continuation

| KMap_exit_body : ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * ('a * 'b) list * ('a, 'c) map * 'a * (('a, 'c) map, 'd * 's, 'r, 'f) continuation -> ('c, 'd * 's, 'r, 'f) continuation

| KView_exit : step_constants * ('a, 's, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KLog : ('a, 's, 'r, 'f) continuation * logger -> ('a, 's, 'r, 'f) continuation

and ('a, 's, 'b, 'f, 'c, 'u) logging_function = ('a, 's, 'b, 'f) kinstr -> Alpha_context.context -> Alpha_context.Script.location -> ('c, 'u) stack_ty -> ('c * 'u) -> unit

and execution_trace = (Alpha_context.Script.location * Alpha_context.Gas.t * (Alpha_context.Script.expr * string option) list) list

and logger = {

log_interp : 'a 's 'b 'f 'c 'u. ('a, 's, 'b, 'f, 'c, 'u) logging_function;
(*
log_interp is called at each call of the internal function interp. interp is called when starting the interpretation of a script and subsequently at each Exec instruction.
*)

log_entry : 'a 's 'b 'f. ('a, 's, 'b, 'f, 'a, 's) logging_function;
(*
log_entry is called before executing each instruction but after gas for this instruction has been successfully consumed.
*)

log_control : 'a 's 'b 'f. ('a, 's, 'b, 'f) continuation -> unit;
(*
log_control is called before the interpretation of the current continuation.
*)

log_exit : 'a 's 'b 'f 'c 'u. ('a, 's, 'b, 'f, 'c, 'u) logging_function;
(*
log_exit is called after executing each instruction.
*)

get_log : unit -> execution_trace option Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t;
(*
get_log allows to obtain an execution trace, if any was produced.
*)

}

and 'ty ty =

| Unit_t : unit ty_metadata -> unit ty

| Int_t : Alpha_context.Script_int.z Alpha_context.Script_int.num ty_metadata -> Alpha_context.Script_int.z Alpha_context.Script_int.num ty

| Nat_t : Alpha_context.Script_int.n Alpha_context.Script_int.num ty_metadata -> Alpha_context.Script_int.n Alpha_context.Script_int.num ty

| Signature_t : Alpha_context.signature ty_metadata -> Alpha_context.signature ty

| String_t : Alpha_context.Script_string.t ty_metadata -> Alpha_context.Script_string.t ty

| Bytes_t : Tezos_protocol_environment_012_Psithaca.Bytes.t ty_metadata -> bytes ty

| Mutez_t : Alpha_context.Tez.t ty_metadata -> Alpha_context.Tez.t ty

| Key_hash_t : Alpha_context.public_key_hash ty_metadata -> Alpha_context.public_key_hash ty

| Key_t : Alpha_context.public_key ty_metadata -> Alpha_context.public_key ty

| Timestamp_t : Alpha_context.Script_timestamp.t ty_metadata -> Alpha_context.Script_timestamp.t ty

| Address_t : address ty_metadata -> address ty

| Bool_t : bool ty_metadata -> bool ty

| Pair_t : ('a ty * Script_ir_annot.field_annot option * Script_ir_annot.var_annot option) * ('b ty * Script_ir_annot.field_annot option * Script_ir_annot.var_annot option) * ('a, 'b) pair ty_metadata -> ('a, 'b) pair ty

| Union_t : ('a ty * Script_ir_annot.field_annot option) * ('b ty * Script_ir_annot.field_annot option) * ('a, 'b) union ty_metadata -> ('a, 'b) union ty

| Lambda_t : 'arg ty * 'ret ty * ('arg, 'ret) lambda ty_metadata -> ('arg, 'ret) lambda ty

| Option_t : 'v ty * 'v option ty_metadata -> 'v option ty

| List_t : 'v ty * 'v boxed_list ty_metadata -> 'v boxed_list ty

| Set_t : 'v comparable_ty * 'v set ty_metadata -> 'v set ty

| Map_t : 'k comparable_ty * 'v ty * ('k, 'v) map ty_metadata -> ('k, 'v) map ty

| Big_map_t : 'k comparable_ty * 'v ty * ('k, 'v) big_map ty_metadata -> ('k, 'v) big_map ty

| Contract_t : 'arg ty * 'arg typed_contract ty_metadata -> 'arg typed_contract ty

| Sapling_transaction_t : Alpha_context.Sapling.Memo_size.t * Alpha_context.Sapling.transaction ty_metadata -> Alpha_context.Sapling.transaction ty

| Sapling_state_t : Alpha_context.Sapling.Memo_size.t * Alpha_context.Sapling.state ty_metadata -> Alpha_context.Sapling.state ty

| Operation_t : operation ty_metadata -> operation ty

| Chain_id_t : Tezos_protocol_environment_012_Psithaca.Chain_id.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Chain_id.t ty

| Never_t : never ty_metadata -> never ty

| Bls12_381_g1_t : Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t ty

| Bls12_381_g2_t : Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t ty

| Bls12_381_fr_t : Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t ty_metadata -> Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t ty

| Ticket_t : 'a comparable_ty * 'a ticket ty_metadata -> 'a ticket ty

| Chest_key_t : Tezos_protocol_environment_012_Psithaca.Timelock.chest_key ty_metadata -> Tezos_protocol_environment_012_Psithaca.Timelock.chest_key ty

| Chest_t : Tezos_protocol_environment_012_Psithaca.Timelock.chest ty_metadata -> Tezos_protocol_environment_012_Psithaca.Timelock.chest ty

and ('top_ty, 'resty) stack_ty =

| Item_t : 'ty ty * ('ty2, 'rest) stack_ty * Script_ir_annot.var_annot option -> ('ty, 'ty2 * 'rest) stack_ty

| Bot_t : (empty_cell, empty_cell) stack_ty

and ('key, 'value) big_map = {

id : Alpha_context.Big_map.Id.t option;

diff : ('key, 'value) big_map_overlay;

key_type : 'key comparable_ty;

value_type : 'value ty;

}

and ('a, 's, 'r, 'f) kdescr = {

kloc : Alpha_context.Script.location;

kbef : ('a, 's) stack_ty;

kaft : ('r, 'f) stack_ty;

kinstr : ('a, 's, 'r, 'f) kinstr;

}

and ('a, 's) kinfo = {

iloc : Alpha_context.Script.location;

kstack_ty : ('a, 's) stack_ty;

}

and (_, _, _, _, _, _, _, _) stack_prefix_preservation_witness =

| KPrefix : ('y, 'u) kinfo * ('c, 'v, 'd, 'w, 'x, 's, 'y, 'u) stack_prefix_preservation_witness -> ('c, 'v, 'd, 'w, 'a, 'x * 's, 'a, 'y * 'u) stack_prefix_preservation_witness

| KRest : ('a, 's, 'b, 'u, 'a, 's, 'b, 'u) stack_prefix_preservation_witness

and ('before, 'after) comb_gadt_witness =

| Comb_one : ('a * ('x * 'before), 'a * ('x * 'before)) comb_gadt_witness

| Comb_succ : ('before, 'b * 'after) comb_gadt_witness -> ('a * 'before, ('a * 'b) * 'after) comb_gadt_witness

and ('before, 'after) uncomb_gadt_witness =

| Uncomb_one : ('rest, 'rest) uncomb_gadt_witness

| Uncomb_succ : ('b * 'before, 'after) uncomb_gadt_witness -> (('a * 'b) * 'before, 'a * 'after) uncomb_gadt_witness

and ('before, 'after) comb_get_gadt_witness =

| Comb_get_zero : ('b, 'b) comb_get_gadt_witness

| Comb_get_one : ('a * 'b, 'a) comb_get_gadt_witness

| Comb_get_plus_two : ('before, 'after) comb_get_gadt_witness -> ('a * 'before, 'after) comb_get_gadt_witness

and ('value, 'before, 'after) comb_set_gadt_witness =

| Comb_set_zero : ('value, _, 'value) comb_set_gadt_witness

| Comb_set_one : ('value, 'hd * 'tl, 'value * 'tl) comb_set_gadt_witness

| Comb_set_plus_two : ('value, 'before, 'after) comb_set_gadt_witness -> ('value, 'a * 'before, 'a * 'after) comb_set_gadt_witness

and (_, _) dup_n_gadt_witness =

| Dup_n_zero : ('a * 'rest, 'a) dup_n_gadt_witness

| Dup_n_succ : ('stack, 'b) dup_n_gadt_witness -> ('a * 'stack, 'b) dup_n_gadt_witness

and ('a, 'b) view_signature =

| View_signature of {

name : Alpha_context.Script_string.t;

input_ty : 'a ty;

output_ty : 'b ty;

}

val kinfo_of_kinstr : ('a, 's, 'b, 'f) kinstr -> ('a, 's) kinfo

type kinstr_rewritek = {

apply : 'b 'u 'r 'f. ('b, 'u, 'r, 'f) kinstr -> ('b, 'u, 'r, 'f) kinstr;

}

val kinstr_rewritek : ('a, 's, 'r, 'f) kinstr -> kinstr_rewritek -> ('a, 's, 'r, 'f) kinstr

val ty_size : 'a ty -> 'a Type_size.t

val comparable_ty_size : 'a comparable_ty -> 'a Type_size.t

val unit_t : annot:Script_ir_annot.type_annot option -> unit ty

val int_t : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_int.z Alpha_context.Script_int.num ty

val nat_t : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_int.n Alpha_context.Script_int.num ty

val signature_t : annot:Script_ir_annot.type_annot option -> Alpha_context.signature ty

val string_t : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_string.t ty

val bytes_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Bytes.t ty

val mutez_t : annot:Script_ir_annot.type_annot option -> Alpha_context.Tez.t ty

val key_hash_t : annot:Script_ir_annot.type_annot option -> Alpha_context.public_key_hash ty

val key_t : annot:Script_ir_annot.type_annot option -> Alpha_context.public_key ty

val timestamp_t : annot:Script_ir_annot.type_annot option -> Alpha_context.Script_timestamp.t ty

val address_t : annot:Script_ir_annot.type_annot option -> address ty

val bool_t : annot:Script_ir_annot.type_annot option -> bool ty

val pair_t : Alpha_context.Script.location -> ('a ty * Script_ir_annot.field_annot option * Script_ir_annot.var_annot option) -> ('b ty * Script_ir_annot.field_annot option * Script_ir_annot.var_annot option) -> annot:Script_ir_annot.type_annot option -> ('a, 'b) pair ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val union_t : Alpha_context.Script.location -> ('a ty * Script_ir_annot.field_annot option) -> ('b ty * Script_ir_annot.field_annot option) -> annot:Script_ir_annot.type_annot option -> ('a, 'b) union ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val union_bytes_bool_t : (Tezos_protocol_environment_012_Psithaca.Bytes.t, bool) union ty

val lambda_t : Alpha_context.Script.location -> 'arg ty -> 'ret ty -> annot:Script_ir_annot.type_annot option -> ('arg, 'ret) lambda ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val option_t : Alpha_context.Script.location -> 'v ty -> annot:Script_ir_annot.type_annot option -> 'v option ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val option_mutez'_t : _ ty_metadata -> Alpha_context.Tez.t option ty

val option_string'_t : _ ty_metadata -> Alpha_context.Script_string.t option ty

val option_bytes'_t : _ ty_metadata -> Tezos_protocol_environment_012_Psithaca.Bytes.t option ty

val option_nat_t : Alpha_context.Script_int.n Alpha_context.Script_int.num option ty

val option_pair_nat_nat_t : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option ty

val option_pair_nat'_nat'_t : _ ty_metadata -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option ty

val option_pair_nat_mutez'_t : _ ty_metadata -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t) pair option ty

val option_pair_mutez'_mutez'_t : _ ty_metadata -> (Alpha_context.Tez.t, Alpha_context.Tez.t) pair option ty

val option_pair_int'_nat_t : _ ty_metadata -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option ty

val option_pair_int_nat'_t : _ ty_metadata -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option ty

val list_t : Alpha_context.Script.location -> 'v ty -> annot:Script_ir_annot.type_annot option -> 'v boxed_list ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val list_operation_t : operation boxed_list ty

val set_t : Alpha_context.Script.location -> 'v comparable_ty -> annot:Script_ir_annot.type_annot option -> 'v set ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val map_t : Alpha_context.Script.location -> 'k comparable_ty -> 'v ty -> annot:Script_ir_annot.type_annot option -> ('k, 'v) map ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val big_map_t : Alpha_context.Script.location -> 'k comparable_ty -> 'v ty -> annot:Script_ir_annot.type_annot option -> ('k, 'v) big_map ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val contract_t : Alpha_context.Script.location -> 'arg ty -> annot:Script_ir_annot.type_annot option -> 'arg typed_contract ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val contract_unit_t : unit typed_contract ty

val sapling_transaction_t : memo_size:Alpha_context.Sapling.Memo_size.t -> annot:Script_ir_annot.type_annot option -> Alpha_context.Sapling.transaction ty

val sapling_state_t : memo_size:Alpha_context.Sapling.Memo_size.t -> annot:Script_ir_annot.type_annot option -> Alpha_context.Sapling.state ty

val operation_t : annot:Script_ir_annot.type_annot option -> operation ty

val chain_id_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Chain_id.t ty

val never_t : annot:Script_ir_annot.type_annot option -> never ty

val bls12_381_g1_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Bls12_381.G1.t ty

val bls12_381_g2_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Bls12_381.G2.t ty

val bls12_381_fr_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Bls12_381.Fr.t ty

val ticket_t : Alpha_context.Script.location -> 'a comparable_ty -> annot:Script_ir_annot.type_annot option -> 'a ticket ty Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

val chest_key_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Timelock.chest_key ty

val chest_t : annot:Script_ir_annot.type_annot option -> Tezos_protocol_environment_012_Psithaca.Timelock.chest ty

type 'a kinstr_traverse = {

apply : 'b 'u 'r 'f. 'a -> ('b, 'u, 'r, 'f) kinstr -> 'a;

}

The following functions named `X_traverse` for X in kinstr, ty, comparable_ty, value provide tail recursive top down traversals over the values of these types.

The traversal goes through a value and rewrites an accumulator along the way starting from some initial value for the accumulator.

All these traversals follow the same recursion scheme: the user-provided function is first called on the toplevel value, then the traversal recurses on the direct subvalues of the same type.

Hence, the user-provided function must only compute the contribution of the value on the accumulator minus the contribution of its subvalues of the same type.

val kinstr_traverse : ('a, 'b, 'c, 'd) kinstr -> 'ret -> 'ret kinstr_traverse -> 'ret

type 'a ty_traverse = {

apply : 't. 'a -> 't ty -> 'a;

apply_comparable : 't. 'a -> 't comparable_ty -> 'a;

}

val comparable_ty_traverse : 'a comparable_ty -> 'r -> 'r ty_traverse -> 'r

val ty_traverse : 'a ty -> 'r -> 'r ty_traverse -> 'r

type 'accu stack_ty_traverse = {

apply : 'ty 's. 'accu -> ('ty, 's) stack_ty -> 'accu;

}

val stack_ty_traverse : ('a, 's) stack_ty -> 'r -> 'r stack_ty_traverse -> 'r

type 'a value_traverse = {

apply : 't. 'a -> 't ty -> 't -> 'a;

apply_comparable : 't. 'a -> 't comparable_ty -> 't -> 'a;

}

val value_traverse : ('t ty, 't comparable_ty) union -> 't -> 'r -> 'r value_traverse -> 'r

val stack_top_ty : ('a, 'b * 's) stack_ty -> 'a ty