Module Tezos_raw_protocol_012_Psithaca.Ticket_balance_key

This module exposes a function for generating a ticket-balance key-hash and an amount, given an owner and a ticket. The key-hash and the amount is used for populating the global ticket-balance table that tracks ownership of different types of tickets.

val ticket_balance_key_and_amount : Alpha_context.context -> owner:Alpha_context.Contract.t -> Ticket_scanner.ex_ticket -> (Alpha_context.Ticket_balance.key_hash * Tezos_protocol_environment_012_Psithaca.Z.t * Alpha_context.context) Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult Tezos_protocol_environment_012_Psithaca.Lwt.t

ticket_balance_key_and_amount ctxt ~owner ex_ticket returns the key_hash of the given owner and ex_ticket, as well as the amount of the ticket.