Module Tezos_raw_protocol_012_Psithaca.Ticket_costs

This module contains constants and utility functions for gas metering functions used for extracting and handling tickets for the global ticket balance table.

module Constants : sig ... end

val consume_gas_steps : Alpha_context.t -> step_cost:Alpha_context.Gas.cost -> num_steps:int -> Alpha_context.t Tezos_protocol_environment_012_Psithaca.Error_monad.tzresult

consume_gas_steps ctxt ~num_steps consumes gas corresponding to a given num_steps and step_cost. It's useful for paying for gas upfront where the number of steps can be determined.

This function is generic and should probably be moved. See issue https://gitlab.com/tezos/tezos/-/issues/1950.

val has_tickets_of_ty_cost : 'a Script_typed_ir.ty -> Saturation_repr.may_saturate Saturation_repr.t

has_tickets_of_ty_cost ty returns the cost of producing a has_tickets, used internally in the Ticket_scanner module.